Restructure Layout + AppRoutes components
