+++ src/react/react-html.jsx
@@ -3,12 +3,15 @@ import { Provider } from 'react-redux';
 import { StaticRouter } from 'react-router';
 
 import AppRoutes from './AppRoutes.jsx';
+import Layout from './Layout.jsx';
 
 export default (request, store) =>
 	renderToString(
 		<Provider store={store}>
 			<StaticRouter location={request.url} context={{}}>
-				<AppRoutes />
+				<Layout>
+					<AppRoutes />
+				</Layout>
 			</StaticRouter>
 		</Provider>
 	);
